Moved all access-rights related away from background_task
