+++ services/catalog/src/simcore_service_catalog/services/access_rights.py
@@ -1,13 +1,101 @@
 """ Services Access Rights policies
 
 """
-from typing import List
+import logging
+from datetime import datetime
+from typing import List, Optional, Tuple
+from urllib.parse import quote_plus
 
+from aiopg.sa.engine import Engine
+from fastapi import FastAPI
 from models_library.services import ServiceAccessRightsAtDB, ServiceDockerData
 from packaging.version import Version
+from pydantic.types import PositiveInt
 
+from ..api.dependencies.director import get_director_api
+from ..db.repositories.groups import GroupsRepository
 from ..db.repositories.services import ServicesRepository
 from ..utils.versioning import as_version, is_patch_release
+
+logger = logging.getLogger(__name__)
+
+OLD_SERVICES_DATE: datetime = datetime(2020, 8, 19)
+
+
+def _is_frontend_service(service: ServiceDockerData) -> bool:
+    return "/frontend/" in service.key
+
+
+async def _is_old_service(app: FastAPI, service: ServiceDockerData) -> bool:
+    # get service build date
+    client = get_director_api(app)
+    data = await client.get(
+        f"/service_extras/{quote_plus(service.key)}/{service.version}"
+    )
+    if not data or "build_date" not in data:
+        return True
+
+    logger.debug("retrieved service extras are %s", data)
+
+    service_build_data = datetime.strptime(data["build_date"], "%Y-%m-%dT%H:%M:%SZ")
+    return service_build_data < OLD_SERVICES_DATE
+
+
+async def create_service_default_access_rights(
+    app: FastAPI, service: ServiceDockerData
+) -> Tuple[Optional[PositiveInt], List[ServiceAccessRightsAtDB]]:
+    """Given a service, it returns the owner's group-id (gid) and a list of access rights following
+    default access-rights policies
+
+    - DEFAULT Access Rights policies:
+        1. All services published in osparc prior 19.08.2020 will be visible to everyone (refered as 'old service').
+        2. Services published after 19.08.2020 will be visible ONLY to his/her owner
+        3. Front-end services are have read-access to everyone
+    """
+    db_engine: Engine = app.state.engine
+
+    groups_repo = GroupsRepository(db_engine)
+    g = await groups_repo.get_everyone_group()
+    everyone_gid = g.gid
+    owner_gid = None
+    reader_gids: List[PositiveInt] = []
+
+    if _is_frontend_service(service) or await _is_old_service(app, service):
+        logger.debug("service %s:%s is old or frontend", service.key, service.version)
+        # let's make that one available to everyone
+        reader_gids.append(everyone_gid)
+
+    # try to find the owner
+    possible_owner_email = [service.contact] + [
+        author.email for author in service.authors
+    ]
+
+    for user_email in possible_owner_email:
+        possible_gid = await groups_repo.get_user_gid_from_email(user_email)
+        if possible_gid:
+            if not owner_gid:
+                owner_gid = possible_gid
+    if not owner_gid:
+        logger.warning("service %s:%s has no owner", service.key, service.version)
+    else:
+        reader_gids.append(owner_gid)
+
+    # we add the owner with full rights, unless it's everyone
+    default_access_rights = [
+        ServiceAccessRightsAtDB(
+            key=service.key,
+            version=service.version,
+            gid=gid,
+            execute_access=True,
+            write_access=(gid == owner_gid),
+            product_name=app.state.settings.access_rights_default_product_name,
+        )
+        for gid in set(reader_gids)
+    ]
+
+    # Patch releases inherit access rights from previous version
+
+    return (owner_gid, default_access_rights)
 
 
 async def evaluate_auto_upgrade_policy(
@@ -21,7 +109,7 @@ async def evaluate_auto_upgrade_policy(
     #
     # SEE https://github.com/ITISFoundation/osparc-simcore/issues/2244)
     #
-    if "/frontend/" in service_metadata.key:
+    if _is_frontend_service(service_metadata):
         return []
 
     service_access_rights = []
